update to isReady()
Previously isReady() would return true if the start of a previous
trigger pull was more than 300ms ago, which meant that it was possible
for isReady() to return true with the flywheels off.

It wouldn't normally matter, because physically the pusher switch can
only be on when the motor switch is closed, but if somehow the switches
were inappropriately aligned it could happen.